Main file to find Discriminant.

+++ SolofQuad.c
@@ -6,13 +6,13 @@
 int main ()
 {
   printf("Enter the coefficients (a,b,c) of quadratic equation:\n");
-  int a,b,c;
-  scanf("%d%d%d",&a,&b,&c);
-  int D;
+  float a,b,c;
+  scanf("%f%f%f",&a,&b,&c);
+  float D;
   D= D1(a,b,c);
   printf("Discriminant of The euation is %d.",D);
-  int rt1=sf1(a,b,c);
-  int rt2=sf2(a,b,c);
+  float rt1=sf1(a,b,c);
+  float rt2=sf2(a,b,c);
   printf("\nroots of the equation are %d %d.",rt1,rt2);
   return 0;
 }
